+++ include/memoria/core/packed/sseq/sseq_fn/pkd_f_sseq_rank_fn.hpp
@@ -24,7 +24,7 @@ namespace memoria {
 template <typename Seq>
 class BitmapRankFn {
 
-    static const int32_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
+    static const size_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
 
     static_assert(BitsPerSymbol == 1, "BitmapRankFn<> can only be used with 1-bit sequences");
 
@@ -33,7 +33,7 @@ class BitmapRankFn {
 public:
     BitmapRankFn(const Seq& seq): seq_(seq) {}
 
-    int32_t operator()(int32_t start, int32_t end, int32_t symbol)
+    size_t operator()(size_t start, size_t end, size_t symbol)
     {
         if (symbol)
         {
@@ -52,7 +52,7 @@ template <typename Seq>
 class SeqRankFn {
     typedef typename Seq::Value                                                 Value;
 
-    static const int32_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
+    static const size_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
 
     static_assert(BitsPerSymbol > 1 && BitsPerSymbol < 8,
             "SeqRankFn<> can only be used with 2-7-bit sequences");
@@ -62,13 +62,13 @@ class SeqRankFn {
 public:
     SeqRankFn(const Seq& seq): seq_(seq) {}
 
-    int32_t operator()(int32_t start, int32_t end, Value symbol)
+    size_t operator()(size_t start, size_t end, Value symbol)
     {
-        int32_t cnt = 0;
+        size_t cnt = 0;
 
         auto buf = seq_.symbols();
 
-        for (int32_t c = start * BitsPerSymbol; c < end * BitsPerSymbol; c += BitsPerSymbol)
+        for (size_t c = start * BitsPerSymbol; c < end * BitsPerSymbol; c += BitsPerSymbol)
         {
             if (TestBits(buf, c, symbol, BitsPerSymbol))
             {
@@ -86,7 +86,7 @@ public:
 template <typename Seq>
 class Seq8RankFn {
 
-    static const int32_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
+    static const size_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
 
     static_assert(BitsPerSymbol == 8,
             "Seq8RankFn<> can only be used with 2-7-bit sequences");
@@ -95,13 +95,13 @@ class Seq8RankFn {
 public:
     Seq8RankFn(const Seq& seq): seq_(seq) {}
 
-    int32_t operator()(int32_t start, int32_t end, int32_t symbol)
+    size_t operator()(size_t start, size_t end, size_t symbol)
     {
-        int32_t cnt = 0;
+        size_t cnt = 0;
 
         auto buf = seq_.symbols();
 
-        for (int32_t c = start; c < end; c++)
+        for (size_t c = start; c < end; c++)
         {
             if (buf[c] == symbol)
             {
